Added React linting to inventory-app and fixed existing violations (#25465)
Adds react and react-hooks linting to the inventory-app example.

This PR is intended to be a first step in adding React linting that
better matches production applications. This in turn will allow us to
write cleaner, more idiomatic React examples. In the future these rules
should be moved to a base package that can be consumed by example
applications.