RuntimeDefinitions: Rename asAlpha to asLegacyAlpha (#25514)
This pull request updates the naming of the function used to cast a
container runtime to its alpha interface, changing it from `asAlpha` to
`asLegacyAlpha`. This change is applied throughout the codebase to
improve clarity and consistency, especially in preparation for future
API changes. All imports, exports, and usages of this function are
updated accordingly in both runtime and test files.