Update driver cache entries to contain the file version  (#25337)
Previously, only snapshot entries for a given document would contain
version information to indicate which version of the document the entry
referred to. This caused a bug where ops from the main document would
appear on previous versions of the document when they should not yet
exist. This PR updates the cache entries of snapshots and ops to both
contain the version information in the cache entry, if it exists.


[AB#47218](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/47218)

Additionally, fixes a few typos I noticed while working in the area.