tree: Fix bug where commits to a branch could be submitted twice (#25520)
## Description

This fixes a bug where `SharedTreeCore.submitCommit` was called twice
when editing a shared branch. An event handler was being registered both
when creating the branch and also when the branch creation was
sequenced. This change adds a callback to `EditManager` which is fired
when a new branch is added, and `SharedTreeCore`'s event handler for
submitting commits is registered as part of that callback.