Add configuredSharedTreeBeta (#25531)
## Description

Add beta version of configuredSharedTreeBeta and the options for
ForestType.

This will allow beta API users to select chunked forest (a prerequisite
for incremental summary), as well as the debug forest. It also sets up
the needed types to stabilize more options.

The alpha versions of these should eventually get renamed, but this does
not do so to reduce the risk of breaks.