fix(api-markdown-documenter): Escape Markdown within HTML (within Markdown) (#25502)
Fixes an issue where HTML content generated within Markdown was not
properly escaping its inner content for Markdown.

Markdown in HTML in Markdown is supported by many Markdown processors.
However, the libraries we use for converting Markdown to HTML assume
that the resulting contents will be used in an HTML context, not a
Markdown one. So, they do not escape the text content as needed for a
Markdown context. This was causing some textual content in multi-line
table cells to not be correctly escaped and therefore parsed as Markdown
content when it would not have been otherwise.

We now explicitly escape content in such cases.