Update to pnpm 10 (#25571)
## Description

Use pnpm 10.

The lock file format seems unchanged except for using larger hashes in
its internal identifiers, which should not cause issues for any tooling.

## Breaking Changes

Developers using corepack will be asked to download pnpm if not already
cached.

Developers using pnpm some other way may have to install it.

Installation scripts for packages are not longer automatically run by
default: only one was found to be needed for CI (puppeteer for
common-utils) but it is possible someone will have a use case which
needs one when reinstalling a package and will have to approve running a
script.