Container:  Connect as read even with pending state (#25601)
This code wasn't correct, and probably isn't the behavior we want.
Specifically, savedOps does not mean there are local changes. savedOps
actually contains all remote ops, and local changes only exist within
the runtimes state.

We want to incorporate staging mode with offline in the near future, and
in that case we may never send local ops if they do exist and user
chooses to disgard them. In that case the user may not expect to edit
the document at all, which is what connecting as write will do.

I think this was an attempt at an optimization, nothing requires write
connect this early, it was just assuming we would edit later, so it
tries to skip the read write transition. As i call out above, this
optimization is based on some false assumptions.