feat(eslint-config-fluid): Promote new rules to errors and replace deprecated `@typescript-eslint` rules (#25451)
# 7.0.0 package update

Promotes the following new rules (previously added as warnings) to
errors:

- `@fluid-internal/fluid/no-hyphen-after-jsdoc-tag`
- `@fluid-internal/fluid/no-file-path-links-in-jsdoc`
- `@fluid-internal/fluid/no-markdown-links-in-jsdoc`
-
[@typescript-eslint/no-empty-object-type](https://typescript-eslint.io/rules/no-empty-object-type/)
-
[@typescript-eslint/no-unsafe-function-type](https://typescript-eslint.io/rules/no-unsafe-function-type/)
-
[@typescript-eslint/no-wrapper-object-types](https://typescript-eslint.io/rules/no-wrapper-object-types/)

## `typescript-eslint` update

`@typescript-eslint/eslint-plugin` and `@typescript-eslint/parser` have
been updated.
These updates includes the deprecation and replacement of a couple of
rules that were included in this library.

The following rules have been deprecated and disabled in this library:

- [ban-types](https://typescript-eslint.io/rules/ban-types/)
-
[no-empty-interface](https://typescript-eslint.io/rules/no-empty-interface/)

They are replaced by these newly promoted rules:

- `no-empty-object-type`
- `no-unsafe-function-type`
- `no-wrapper-object-types`

## Misc. notes

Note: Biome's config, which our `strict-biome` config extends, disables
`no-empty-interface`, but does not disable `no-empty-object-type`. As a
result, our `strict-biome` config now enables the latter rule when the
former was not enabled.

This PR also enables the
[no-new-native-nonconstructor](https://eslint.org/docs/latest/rules/no-new-native-nonconstructor)
rule. This rule would otherwise have been enabled but was disabled by
the typescript-eslint configuration. That configuration no longer
disables this rule.

This PR also pre-applies the appropriate rule disables where appropriate
in the code. The existing disables of the deprecated rules may be
removed _after_ the repo has been updated to consume this new version of
this package.