Cleanup tree-agent (#25612)
* Remove stateful branch member in favor of functional patterns
* Simplify branching flow. Branching now happens at the start of each
query, and once again for each edit within the query. Previously,
branching happened upon the creation of the agent, and in a couple of
places during the query/edit process, and required more rebasing to stay
in sync.
* Various cleanups to code and LLM messaging flow - e.g. we no longer
throw an error if we exceed the maximum number of edits, but instead
return a message with instructions for the user.
* Moved `treeToString` inside of logger object
* Properly set the outer branch dirty bit