Allow build scripts already in use by client (#25617)
## Description

With CI passing without these, I thought we were fine, but it seems like
linting depends on at least one of them.

As we used to run when we were on pnpm 9 automatically, enabling them on
pnpm 10 is not much of a regression, but the set we do run should be
audited and trimmed down to what's required in the future.

If getting strange errors when linting locally, removing your
`./node_modules` files and running `pnpm i` after this change may help.

List generated using `pnpm approve-builds` and selecting all.