Add model injection to tree-agent (#25624)
This restructures the tree-agent entry point to be supplied with a
"model" interface that coordinates API calls to an LLM service.
Previously, the library was hardcoded to use the langchain library,
which means that any services not implemented as a langchain plugin were
not usable.

This also adds several unit tests of high-level agent functionality
using a mock model.

This architecture is not final and will likely be improved going
forward, but is important in the short term to remove the dependency on
langchain.