Simplify tool registration and various other improvements to tree-agent (#25634)
The SharedTreeChatModel no longer supplies an edit function name, and
instead must supply a tool name (or not, in which case editing will be
disabled). Documentation has been improved as well.

* treeToString has been removed. It was not particularly useful in logs
(in fact, it was sometimes confusing), and it caused a lot of generic
parameters and complicated type intersections (especially when using the
subtree API).
* Added a test for scoping a SharedTreeSemanticAgent to a subtree
* Added documentation for more APIs