Handle arbitrary edit function names in tree-agent (#25636)
In practice, LLMs will likely continue to generate functions of the same
form and name `function editName(...` because that's what we provide in
the example. However, now, if they do it slightly differently once in a
while, we will still tolerate it. This uses the parsing library Acorn to
find the function name robustly.

This also removes the `toEditTreeCode` test helper and inlines the edit
function code instead - it was harder to read with the helper,
ironically. Additionally, SharedTreeSemanticAgent is now properly
exported as non-type.