SharedTree: Refactor annotated types API (#25595)
## Description

The current API for annotated allowed types has some issues, mainly:
1. it causes a divergence between the new alpha types and existing types
which requires unannotation
2. it can be impossible to parse ImplicitAnnotatedAllowedTypes without
inspecting the type too closely which causes issues with recursive types
3. We are missing APIs for actually constructing the sealed types so
there isn't a supported way to use these APIs.

This change solves these by introducing a new type `AllowedTypesFull`
with is a sealed type (so we can add helpful properties and methods to
it later) that stores the annotations, but also implements
`AllowedTypes` (Using proxy since it's an array).

This restores the ability to normalize all allowed types representations
to a single type which can be used in all cases. This also makes it
possible to use the existing recursive type logic with annotated types,
implicitly solving that problem as well.

This change effectively removes complexity from a lot of places (all the
unannotation logic, and handling multiple representations) and
centralizes it in one class. This seems to be an overall simplification
and also be more forwards compatible with future schema additions that
the current approach.

## Breaking Changes

Many annotated types alpha APIs have been impacted, and there should be
good alternatives for all of them, see the changeset.