Add sandboxing plugin to tree-agent (#25645)
* Add plugin hook for evaluating LLM-generated JS code. This defaults to
a simple eval, but can be overridden by the user to do something more
secure.
* Add drop-in implementation of the above hook that uses SES to secure
the execution environment
* Update signature, docs and implementation pattern for the _validation_
hook to match that of the new evaluation hook.
* Changed generated code scheme from "generate a function with the given
context parameter" to "generate a code snippet that has access to a
context variable". This is easier to implement, especially given the new
sandbox capability. The system prompt was necessarily updated to reflect
this scheme. Manual testing has shown that there is no observable
difference in the quality/correctness of the LLM-generated code under
this scheme.
* The new generated code scheme removes the need to parse the JS in
order to find the name of the function, therefore, the Acorn library
dependency and all associated helpers and tests were removed.

----

Why a plugin model? Why not just sandbox by default? The primary options
for sandboxing in a browser environment are:

1. Isolate to a webworker/service
2. Isolate to an iframe
3. Use a library like SES to "lock down" the current environment

Options 1 and 2 require a different architecture entirely, because data
must be fully serialized to pass between the agent and the generated
code. Therefore, option 3 was chosen as the first attempt at making
tree-agent resistant to attacks from malicious generated code.

When SES locks down the JS environment, it does so by mutating various
aspects of the global scope (e.g. freezing built-in class prototypes).
This affects the entire runtime. If this were the default behavior, then
it would mean that using tree-agent would cause these side effects -
which could affect other unrelated parts of an application - and that
may not be what the application author wants. So, a lazy loaded plugin
model was chosen here instead so that an app author can choose what to
do. Providing the drop-in SES evaluator for them already makes it very
easy to use if that's what they want.