(tree) Update chunking algorithm to use polymorphic shape for incremental fields (#25400)
Currently, fields that are specified as incremental via the
`shouldEncodeIncrementally` callback can be monomorphic causing them to
be combined in a uniform chunk with its parent.
This will make the field to not be incrementally summarized - When
incremental summary builder calls `ChunkedForest::chunkField` for such
fields, it will return a new chunk every time causing it to encode the
chunk again and not re-using a summary handle for it.

This PR changes the chunking algorithm so that it always uses
polymorphic shape for nodes / fields that are specified as incremental.
This is done by plumbing the `shouldEncodeIncrementally` function to the
`tryShapeFromSchema` function which decides the shapes for fields and
nodes. If `shouldEncodeIncrementally` returns true, a polymorphic shape
is returned.


[AB#41866](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/41866)