Replacing Lerna with pnpm on LTS (#25537)
This PR replaces lerna for client packages in the LTS branch with pnpm,
and merges azure into the 'client' package group

The build tools version used by LTS isn't compatible with pnpm's
workspace:~ dependency protocol, so we had to use a version of
build-tools from main which does support that protocol. That introduced
a few ancilliary changes we had to make:

1 - build-tools from main requires node>=20. As a result we've updated
LTS to use node 20.15.1 (the min required version on main)

2 - Updating to node 20.15.1 required us to update our eslint dependency
to a version compatible with node 20
2a - Updating eslint also introduced a number of linter changes to
various code files

3 - the latest version of build-tools from main had similar but
different type test generation code - we had to update the
typeValidation JSON in various package.jsons, and the build-tools
generated type tests are generated with slightly different names + paths

4 - we disabled the layer-checking policy check on LTS (per offline
discussion, we decided it was unlikely there would be major dependency
changes to LTS in the future, and so it would be a lot of work for
little payoff to enable layer-checking using the latest build-tools
version on LTS)

5 - Since main considers azure packages to be part of the client
workgroup, we've added azure packages to this change and made them use
pnpm
5a - This also lets us remove the build-azure pipeline, since azure gets
included with the client packages build

6 - Several pipelines were updated to use pnpm, install pnpm
dependencies, and use flub to publish packages. Where possible, the
implementations have been pulled directly from main

7 - build-tools sources (and associated pipeline to build build-tools)
have been removed. We're now using build-tools from main, which means
this code is no longer needed to be built or maintained (NB: the server/
code which uses old build-tools is planned to be removed in a subsequent
PR)


Workarounds needed:
1) api-extractor dependencies are using ~7.22.2 instead of ^7.22.2 now.
This was to work around an issue with api-extractor, where ^
dependencies used an inconsistent version of the typescript compiler
that caused build failures (and the command line switch to point
api-extractor to the right compiler to use did not work)

2) eslint configurations have been set to use 0.28.2000. This resulted
in a downgraded eslint config for some projects. This was done because
of the migration to node 20 - eslint config 0.28.2000's dependencies
were compatible with node 20, but 2.0.0 included extra dependencies
which were incompatible with node 20

3) @types/node set to 14.18.0 everywhere - trying to use a different
types/node version caused typescript compilation failures when
processing d.ts files of those libraries, and options like skipLibCheck
etc did not fix it. Setting all projects to @types/node 14.18.0 resolved
the issue (NB: main has a similar workaround of using node types v 18
due to compilation issues, despite it using node 20+)