(compat) Fix incorrect usages of storage service at Runtime and Loader layer (#25722)
There are a couple of incorrect usages of the storage service at the
Runtime and Loader layer after `IContainerStorageService` and
`IRuntimeStorageService` were added by
https://github.com/microsoft/FluidFramework/pull/25057. This PR fixes
them.

Also, removed a test that was ensuring `getSnapshotTree` cannot be
accessed from data store context. With
https://github.com/microsoft/FluidFramework/pull/25057,
`getSnapshotTree` is not accesible from data store context's storage
anymore so this test is not needed.