Add timestamps at message push time in attributor tests (#25748)
## Description

Adjust the way that SharedString attribution tests inject timestamps
into messages to be a bit more realistic; rather than at op event
raising time, we do so "on the server" when the messages are "sent".
This necessitates updating the snapshots to reflect that catchup ops
have timestamps (as they would with a real server implementation).

Motivation: #25745 adjusts how op events are fired in a way that breaks
the mock implementation here.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>